third_party/WebKit/LayoutTests/inspector/console/console-worker-nested-imports-syntax-error.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>
function startWorker()
{
    var worker = new Worker("resources/importScripts-1.js");
}

function test()
{
    InspectorTest.waitForConsoleMessages(1, step1);
    InspectorTest.evaluateInPage("startWorker();");

    function step1()
    {
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>
Tests that nested import scripts in worker show correct stack on syntax error.
</p>
</body>
</html>
