third_party/WebKit/LayoutTests/inspector/console/console-worker-nested-imports-syntax-error-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected identifier
Tests that nested import scripts in worker show correct stack on syntax error.

Message count: 1
invalidScript.js:1 Uncaught SyntaxError: Unexpected identifier(…)(anonymous function) @ importScripts-3.js:1foo @ importScripts-2.js:3(anonymous function) @ importScripts-2.js:5(anonymous function) @ importScripts-1.js:1
